+++ src/app/modules/slormancer/model/content/enum/mechanic-type.ts
@@ -8,4 +8,6 @@ export enum MechanicType {
     Dart = 'dart',
     Frostbolt = 'frostbolt',
     LightningRod = 'rod',
+    SoulBound = 'soul_bound',
+    Burn = 'burn',
 }