+++ src/app/modules/slormancer/services/slormancer-character.updater.service.ts
@@ -268,6 +268,15 @@ export class SlormancerCharacterUpdaterService {
                 if (stats.extractedStats['primary_secondary_skill_lightning_imbued'] !== undefined) {
                     skillAndUpgrades.skill.elements.push(SkillElement.Lightning);
                 }
+                if (stats.extractedStats['primary_secondary_skill_light_imbued'] !== undefined) {
+                    skillAndUpgrades.skill.elements.push(SkillElement.Light);
+                }
+                if (stats.extractedStats['primary_secondary_skill_shadow_imbued'] !== undefined) {
+                    skillAndUpgrades.skill.elements.push(SkillElement.Shadow);
+                }
+                if (stats.extractedStats['primary_secondary_skill_fire_imbued'] !== undefined) {
+                    skillAndUpgrades.skill.elements.push(SkillElement.Fire);
+                }
             }
         }
     }